removed allotment gui and added click to add
