+++ client/components/index.js
@@ -11,7 +11,6 @@ export { default as Room } from './room';
 export { default as NewGame } from './new-game';
 export { default as Sidebar } from './sidebar';
 export { default as Settings } from './settings';
-export { default as AllotmentGUI } from './allotment';
 export { default as CombatRisk } from './combat-risk';
 export { default as CombatCustom } from './combat-custom';
 export { default as ChannelList } from './channelList';
